Rename common/warn.ts to common/punishments.ts
